# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Tracker API
  description: Full REST API for managing portfolio data — accounts, assets, snapshots, positions, income, expenses, and allocation targets.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/accounts:
    get:
      summary: List all accounts
      tags: [Accounts]
      responses:
        '200':
          description: Array of accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Account' }
    post:
      summary: Create account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
...
  /api/v1/assets:
    get:
      summary: List all assets
      tags: [Assets]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Asset' }
    post:
      summary: Create asset
      tags: [Assets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetInput' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
...
  /api/v1/positions:
    get:
      summary: List all positions
      tags: [Positions]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Position' }
    post:
      summary: Create or update position (upsert)
      tags: [Positions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PositionInput' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
...
  /api/v1/income:
    get:
      summary: List all income records
      tags: [Income]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Income' }
    post:
      summary: Create income record
      tags: [Income]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncomeInput' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Income' }
...
  /api/v1/expenses:
    get:
      summary: List all expense records
      tags: [Expenses]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Expense' }
    post:
      summary: Create expense record
      tags: [Expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpenseInput' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
...
  /api/v1/targets:
    get:
      summary: List allocation targets
      tags: [Targets]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Target' }
    post:
      summary: Create or upsert allocation target
      tags: [Targets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TargetInput' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Target' }
...
components:
  schemas:
...
    Account:
      type: object
      description: "A financial account — a brokerage, bank, exchange, or physical asset container. Each account can hold multiple assets."
      properties:
        id:
          type: integer
          description: "Auto-generated unique identifier"
          example: 2
        name:
          type: string
          description: "Display name for this account. Must be unique."
          example: "MyBroker"
        type_code:
          type: string
          description: |
            Account classification:
            - `broker` — Stock brokerage
            - `bank` — Traditional bank
            - `exchange` — Crypto exchange
            - `cash` — Physical cash or misc cash-like account
            - `crypto` — Aggregated crypto account (single bucket for all on-chain holdings)
            - `asset` — Physical asset like a car
          enum: [broker, bank, exchange, cash, crypto, asset]
          example: "broker"
        institution:
          type: string
          nullable: true
          description: "The company/institution behind this account. Used for display only."
          example: "Example Brokerage Inc."
        chain_code:
          type: string
          nullable: true
          description: |
            Blockchain network this account operates on (only for crypto/on-chain accounts).
            Values: `bitcoin`, `ethereum`, `solana`, `cosmos`, `kujira`. Leave null for traditional finance accounts.
          example: "solana"
        active:
          type: integer
          description: "1 = active, 0 = soft-deleted (hidden from UI but preserved in history)"
          example: 1
        notes:
          type: string
          nullable: true
          description: "Free-text notes"
          example: "Main US brokerage"
    AccountInput:
      type: object
      description: "Fields for creating or updating an account"
      required: [name, type_code]
      properties:
        name:
          type: string
          description: "Unique display name"
          example: "MyExchange"
        type_code:
          type: string
          description: "Account type: broker, bank, exchange, cash, crypto, or asset"
          enum: [broker, bank, exchange, cash, crypto, asset]
          example: "exchange"
        institution:
          type: string
          description: "Company name (optional)"
          example: "Example Exchange Holdings"
        chain_code:
          type: string
          description: "Blockchain network (optional, e.g. solana, ethereum)"
          example: null
        notes:
          type: string
          description: "Free-text notes (optional)"
          example: "Spot + futures"
    Asset:
      type: object
      description: "A tradeable asset — a stock ticker, crypto token, fiat currency, stablecoin, or NFT."
      properties:
        id:
          type: integer
          example: 15
        symbol:
          type: string
          description: "Ticker/token symbol. Unique within its type_code."
          example: "BTC"
        name:
          type: string
          nullable: true
          description: "Human-readable full name"
          example: "Bitcoin"
        type_code:
          type: string
          description: |
            Asset classification:
            - `crypto` — Cryptocurrency token (BTC, ETH, SOL, JUP, KMNO...)
            - `stock` — Stock or ETF
            - `stable` — Stablecoin (USDC)
            - `fiat` — Fiat currency (USD, EUR, GBP, PYG)
            - `nft` — Non-fungible token
          enum: [crypto, stock, stable, fiat, nft]
          example: "crypto"
        chain_code:
          type: string
          nullable: true
          description: "Blockchain this asset lives on (for crypto only). Values: bitcoin, ethereum, solana, cosmos"
          example: "bitcoin"
        risk_code:
          type: string
          nullable: true
          description: |
            Risk classification (for crypto):
            - `cat1_safe` — Low risk, blue chip (BTC, ETH, SOL)
            - `cat2_medium` — Medium risk (KMNO, PYTH, PRCL, STX)
            - `cat3_high` — High risk, speculative (JUP, BONK, LUNA)
            Leave null for non-crypto assets.
          enum: [cat1_safe, cat2_medium, cat3_high]
          example: "cat1_safe"
        coingecko_id:
          type: string
          nullable: true
          description: "CoinGecko API identifier. Used by the pricefeed microservice to fetch prices. Find IDs at coingecko.com/en/api."
          example: "bitcoin"
        yahoo_ticker:
          type: string
          nullable: true
          description: "Yahoo Finance ticker. Used by the pricefeed for stock prices."
          example: "VOO"
        target_pct:
          type: number
          nullable: true
          description: "Target allocation as a decimal. 0.70 = 70% of its category. Used in the crypto actual-vs-target chart."
          example: 0.70
        is_stable:
          type: integer
          description: "1 = stablecoin (USDC), 0 = normal asset"
          example: 0
        active:
          type: integer
          description: "1 = active, 0 = deactivated"
          example: 1
    AssetInput:
      type: object
      description: "Fields for creating or updating an asset"
      required: [symbol, type_code]
      properties:
        symbol:
          type: string
          example: "SOL"
        name:
          type: string
          example: "Solana"
        type_code:
          type: string
          enum: [crypto, stock, stable, fiat, nft]
          example: "crypto"
        chain_code:
          type: string
          description: "Blockchain (e.g. solana, ethereum, bitcoin). Null for stocks/fiat."
          example: "solana"
        risk_code:
          type: string
          description: "Risk level: cat1_safe, cat2_medium, cat3_high, or empty"
          enum: ["", cat1_safe, cat2_medium, cat3_high]
          example: "cat1_safe"
        coingecko_id:
          type: string
          description: "CoinGecko ID for price fetching (e.g. 'solana', 'bitcoin', 'ethereum')"
          example: "solana"
        yahoo_ticker:
          type: string
          description: "Yahoo Finance ticker for stock prices (e.g. 'VOO', 'AAPL')"
          example: null
        target_pct:
          type: number
          description: "Target allocation as decimal (0.30 = 30%). Null if no target."
          example: 0.30
        is_stable:
          type: boolean
          description: "true for stablecoins (USDC), false for everything else"
          example: false
...
    Position:
      type: object
      description: |
        Current holding — what you own RIGHT NOW. Unlike snapshots (historical time-series),
        positions represent the latest state. Updated by the pricefeed or manually.
        Primary key is (account_id, asset_id) — one position per account+asset pair.
      properties:
        account_id:
          type: integer
          description: "Account that holds this position"
          example: 2
        asset_id:
          type: integer
          description: "Asset being held"
          example: 9
        quantity:
          type: number
          description: "Number of shares/tokens/units"
          example: 399.72
        avg_cost:
          type: number
          nullable: true
          description: "Average cost per unit (cost basis). Used for P&L calculation."
          example: 540.00
        last_price:
          type: number
          nullable: true
          description: "Most recent price from pricefeed. 0 if pricefeed hasn't run yet."
          example: 605.00
        value_usd:
          type: number
          description: "Current total value (quantity x last_price)"
          example: 241918.60
        as_of:
          type: string
          description: "Date of the last price update"
          example: "2026-04-25"
    PositionInput:
      type: object
      description: "Create or update a position. Upserts on (account_id, asset_id)."
      required: [account_id, asset_id, quantity, value_usd]
      properties:
        account_id:
          type: integer
          description: "Account ID"
          example: 2
        asset_id:
          type: integer
          description: "Asset ID"
          example: 9
        quantity:
          type: number
          description: "Number of shares/tokens"
          example: 400.00
        avg_cost:
          type: number
          description: "Average purchase price per unit"
          example: 540.00
        last_price:
          type: number
          description: "Current market price per unit (optional, updated by pricefeed)"
          example: 605.00
        value_usd:
          type: number
          description: "Total current value"
          example: 242000.00
    Income:
      type: object
      description: "Monthly income record tracking salary, bonus, taxes, and employer."
      properties:
        id:
          type: integer
          example: 74
        as_of:
          type: string
          format: date
          description: "Month this income is for (typically end-of-month date like 2026-04-30)"
          example: "2026-04-30"
        salary_usd:
          type: number
          description: "Monthly gross salary in USD"
          example: 8000.00
        per_year_usd:
          type: number
          description: "Annual salary (informational, from xlsx import)"
          example: 96000.00
        bonus_usd:
          type: number
          description: "Bonus received this month"
          example: 0.00
        taxes_usd:
          type: number
          description: "Taxes withheld this month"
          example: 2000.00
        company:
          type: string
          nullable: true
          description: "Employer name"
          example: "Acme Corp"
    IncomeInput:
      type: object
      description: "Create or update an income record. Upserts on as_of date."
      required: [as_of, salary_usd]
      properties:
        as_of:
          type: string
          format: date
          description: "Month date (YYYY-MM-DD, typically end of month)"
          example: "2026-04-30"
        salary_usd:
          type: number
          description: "Gross monthly salary"
          example: 8000.00
        bonus_usd:
          type: number
          description: "Monthly bonus (default 0)"
          example: 0.00
        taxes_usd:
          type: number
          description: "Taxes withheld (default 0)"
          example: 2000.00
        company:
          type: string
          description: "Employer name"
          example: "Acme Corp"
    Expense:
      type: object
      description: "Expense record with optional category and labels."
      properties:
        id: { type: integer, example: 73 }
        as_of: { type: string, format: date, example: "2026-04-30" }
        amount_usd: { type: number, example: 42.50 }
        place: { type: string, nullable: true, example: "Sushi-X" }
        notes: { type: string, nullable: true, example: null }
        category_id: { type: integer, nullable: true, example: 4 }
        category_name: { type: string, nullable: true, description: "Joined from categories.name", example: "Restaurants" }
        labels:
          type: array
          description: "Labels applied to this expense"
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
    ExpenseInput:
      type: object
      description: |
        Create or update an expense. Upserts on as_of date. Pass `label_ids` to replace the
        full set of labels on this expense (omit to leave existing labels untouched on update).
      required: [as_of, amount_usd]
      properties:
        as_of: { type: string, format: date, example: "2026-04-26" }
        amount_usd: { type: number, example: 42.50 }
        place: { type: string, example: "Sushi-X" }
        category_id: { type: integer, nullable: true, example: 4 }
        label_ids:
          type: array
          items: { type: integer }
          description: "Replaces existing labels on save. Send [] to clear."
          example: [1, 2]
...
    TargetInput:
      type: object
      description: "Create or update an allocation target. Upserts on (category, market_mode)."
      required: [category, target_pct]
      properties:
        category:
          type: string
          description: "Category name (stocks, stable_yielding, crypto, cash, car)"
          example: "stocks"
        market_mode:
          type: string
          enum: [bull, crab, bear]
          description: "Market regime. Defaults to 'crab' if omitted."
          example: "bull"
        target_pct:
          type: number
          description: "Target as decimal (0.45 = 45%)"
          example: 0.45
        notes:
          type: string
          description: "Optional notes"
          example: "Large cap index funds"
...